pyproject.toml
[project]
name = "pydentic-new"
version = "0.1.0"
description = "Content Creation App"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.126.0",
    "firecrawl-py>=2.8.0",
    "langchain-anthropic>=0.3.15",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.2",
    "langchain-mistralai>=0.2.10",
    "langchain-openai>=0.3.21",
    "langchain-perplexity>=0.1.1",
    "markdown2>=2.5.3",
    "pdfkit>=1.0.0",
    "pydantic>=2.11.5",
    "python-pptx>=1.0.2",
    "streamlit>=1.45.1",
]
